Fix month format and improve loading: wait for iPhone frame and wallpaper
- Change month format from 'January' to 'Jan' in visit names
- Add loading state to wait for iPhone frame image and wallpaper
- Show blue background until both images are loaded
- Prevents iPhone frame from appearing after content

+++ api/notion/track-visit.ts
@@ -80,7 +80,7 @@ export default async function handler(
 
     // Generate name: "Visit {number} | {Month} {Year}"
     const now = new Date();
-    const monthYear = now.toLocaleDateString('en-US', { month: 'long', year: 'numeric' });
+    const monthYear = now.toLocaleDateString('en-US', { month: 'short', year: 'numeric' });
     const visitNumber = visitCount + 1;
     const entryName = `Visit ${visitNumber} | ${monthYear}`;
